Add files via upload
The IPL match score predictor ML-model uses a large dataset of historical cricket match data to train itself and learn patterns and correlations between various factors and the outcome of matches. The model uses a variety of algorithms, such as regression analysis, decision trees, and random forest, to make accurate predictions.

The model can be used by cricket enthusiasts, fans, and bookmakers to predict the score of a match, the likelihood of a team winning, and even the most valuable player of a match. The model can be trained and updated regularly to keep up with changes in team compositions and player performance, making it a valuable tool for predicting IPL match outcomes.

Overall, the IPL match score predictor ML-model is a powerful tool for predicting the outcome of IPL matches, and it has the potential to revolutionize the way fans and bookmakers make predictions and place bets.